requirements.txt
streamlit
openai
python-dotenv
semantic-kernel
pyyaml 